refactor: move manifest to utils/api
