+++ src/lib/utils/api/index.ts
@@ -2,3 +2,4 @@ export * from './auth';
 export * from './delete';
 export * from './health';
 export * from './registry';
+export * from './manifest';
