+++ src/lib/utils/formatting/size.ts
@@ -1,0 +1,9 @@
+export function formatSize(bytes: number): string {
+	const units = ['B', 'KB', 'MB', 'GB', 'TB'];
+	let i = 0;
+	while (bytes >= 1024 && i < units.length - 1) {
+		bytes /= 1024;
+		i++;
+	}
+	return `${bytes.toFixed(2)} ${units[i]}`;
+}
